+++ src/components/AdminPanel/common/TokensPoolControl.tsx
@@ -4,7 +4,6 @@ import { useCallback, useEffect, useMemo, useState } from 'react';
 import { BigNumber } from 'ethers';
 
 import { Button } from '@/components/ui/Button/Button';
-import { useAccessControl } from '@/hooks/admin/useAccessControl';
 import { useContractsAddresses } from '@/hooks/admin/useContractsAddresses';
 import { useTokensPoolControl } from '@/hooks/admin/useTokensPoolControl';
 import { ContractsEnum } from '@/hooks/contracts/useContractAbi';
@@ -26,7 +25,6 @@ export const TokensPoolControl = ({
   contractName: TokensPoolTypes;
 }) => {
   const tokensPool = useTokensPoolControl(contractName);
-  const tokensPoolAccessControl = useAccessControl(contractName);
 
   const {
     isOpen: isOpenAddCategory,
@@ -46,13 +44,13 @@ export const TokensPoolControl = ({
       </Flex>
 
       <AddressesListControl
-        addresses={tokensPoolAccessControl.adminsRequest.data || []}
+        addresses={tokensPool.adminsRequest.data || []}
         label="Edit admin"
         listLabel="Admins"
         addActionLabel="Grant role"
         removeActionLabel="Revoke role"
-        onAdd={tokensPoolAccessControl.grantAdminRole.mutateAsync}
-        onRemove={tokensPoolAccessControl.revokeAdminRole.mutateAsync}
+        onAdd={tokensPool.grantAdminRole.mutateAsync}
+        onRemove={tokensPool.revokeAdminRole.mutateAsync}
       />
 
       <Text fontWeight="500" my="8px" fontSize="18px">
