+++ src/components/Raffle/RaffleDescrption.tsx
@@ -25,11 +25,6 @@ export const RaffleDescription: FC<RaffleDescriptionProps> = ({
     [winnersForLevel]
   );
 
-  const formattedDescription = useMemo(
-    () => description?.replace(/(?<=\d) /g, '\xa0'),
-    [description]
-  );
-
   return (
     <Box
       bgColor="bgGreen.50"
@@ -118,7 +113,7 @@ export const RaffleDescription: FC<RaffleDescriptionProps> = ({
 
         {description ? (
           <Text textStyle="text1" mt={{ sm: '18px', '2xl': '26px' }}>
-            {formattedDescription}
+            {description}
           </Text>
         ) : null}
       </Flex>
